requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.3.0
et-xmlfile==1.1.0
Flask==2.2.4
Flask-Mail==0.9.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.2.2
